print frq: Valid values of vector of NA's are 0.
To allow to show the frequency of a vector with only NA values, the valid N should be 0. In order to get that, since the data frame of frequencies consists of only 1 row which cannot be computed, and row 0 is always missing, count of valid frequencies should start by row 0.